+++ CHANGES.rst
@@ -15,6 +15,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- #406: Allow access to the country API via empty GET requests.
+
 - Use custom JSON encoding for kombu transport to support date/time objects.
 
 - Change MEASURE_TYPE_CODE, SCORE_TYPE, STAT_TYPE into IntEnums.